+++ tortoise/backends/sqlite/schema_generator.py
@@ -1,8 +1,7 @@
 from typing import Any, List
 
-from pymysql.converters import encoders
-
 from tortoise.backends.base.schema_generator import BaseSchemaGenerator
+from tortoise.converters import encoders
 
 
 class SqliteSchemaGenerator(BaseSchemaGenerator):
@@ -38,4 +37,4 @@ class SqliteSchemaGenerator(BaseSchemaGenerator):
         return default_str
 
     def _escape_default_value(self, default: Any):
-        return encoders.get(type(default))(default)
+        return encoders.get(type(default))(default)  # type: ignore